address requested changes
